Seperated custom form into a seperate file.
